Move additional methods in SuppressWarningsSubprocessorCore to jdt.core.manipulation
